# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Coffee Website API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/cart/{id}": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "نمایش سبد خرید یک کاربر",
                "description": "نمایش آیتم‌ها و محصولات موجود در سبد خرید بر اساس آیدی کاربر",
                "operationId": "7281202a117e3e9c5957e17acb401141",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "آیدی کاربر",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "نمایش سبد خرید یا پیام خالی بودن آن",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "product": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 10
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "گوشی سامسونگ A73"
                                                                    },
                                                                    "price": {
                                                                        "type": "number",
                                                                        "format": "float",
                                                                        "example": 14500000
                                                                    },
                                                                    "image": {
                                                                        "type": "string",
                                                                        "example": "a73.jpg"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            },
                                                            "quantity": {
                                                                "type": "integer",
                                                                "example": 2
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "سبد خرید شما خالی است."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/cart/add": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "افزودن محصول به سبد خرید کاربر",
                "description": "اگر محصول قبلاً داخل سبد بوده، فقط تعدادش افزایش می‌یابد. در غیر این صورت، محصول جدید به سبد اضافه می‌شود.",
                "operationId": "00b66d86de1d25d51dd2bc104d5116aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "userid",
                                    "productid",
                                    "quantity"
                                ],
                                "properties": {
                                    "userid": {
                                        "description": "آیدی کاربر",
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "productid": {
                                        "description": "آیدی محصول",
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "quantity": {
                                        "description": "تعداد",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "محصول با موفقیت به سبد خرید اضافه شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product added to cart"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "خطای اعتبارسنجی",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/cart/remove/{userid}/{productid}/{quantity}": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "حذف یا کاهش تعداد یک محصول از سبد خرید کاربر",
                "description": "اگر تعداد جدید صفر یا کمتر باشد، محصول از سبد خرید حذف می‌شود. در غیر این صورت فقط مقدار آن کاهش می‌یابد.",
                "operationId": "aabfa069eac80f2e1baca1ff04abce7e",
                "parameters": [
                    {
                        "name": "userid",
                        "in": "path",
                        "description": "آیدی کاربر",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    },
                    {
                        "name": "productid",
                        "in": "path",
                        "description": "آیدی محصولی که باید از سبد خرید حذف یا کم شود",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 12
                        }
                    },
                    {
                        "name": "quantity",
                        "in": "path",
                        "description": "تعداد برای کم‌کردن (مثبت یا منفی). اگر نتیجه نهایی صفر یا کمتر شود، آیتم حذف می‌شود.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": -1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "محصول حذف شد یا تعداد به‌روزرسانی شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "محصول از سبد خرید حذف شد"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "تعداد به‌روزرسانی شد"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "مقدار وارد شده معتبر نیست",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "تعداد وارد شده معتبر نیست"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "دسترسی غیرمجاز به آیتم",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "دسترسی غیرمجاز"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "آیتم مورد نظر یافت نشد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "محصول مورد نظر در سبد خرید پیدا نشد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "خطای ناشناخته",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "یک خطای ناشناخته رخ داد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}